emoji: take into account change in  `emojify-completing-read`
`emojify-completing-read` used to take as input a predicate that takes
two arguments as input.

We seem to take the second argument containing something like
`:flag-yt:` and we used it to look at `slack-emoji-master` to see if
the emoji exists.

Now I takes a single argument which looks like:

`":flag-yt: - Mayotte (github)`

This does not match the contents of the cache anymore.

In this patch, I propose changing the format of the cache, to match
the cache created by `emojify`, and then changing the lambda passed to
`emojify-completing-read` to make take into account the new format of
the list.

This patch is not complete yet, as we need to make sure `(oref team
emoji-master)` is built with the same format.